+++ infra/database/sql/common.sql
@@ -1,4 +1,5 @@
-CREATE USER connect_admin WITH PASSWORD 'passpass';
+-- CREATEROLE: the analytics migration creates the role analytics_reader.
+CREATE USER connect_admin WITH PASSWORD 'passpass' CREATEROLE;
 CREATE DATABASE "connect" OWNER connect_admin;
 CREATE DATABASE connect_test OWNER connect_admin;
 \connect connect